refactor(core): 拆分渠道配置与消息状态服务
