+++ lib/models/channel_config_department.dart
@@ -1,0 +1,247 @@
+part of 'channel_config.dart';
+
+// ==================== 职能部门渠道 ====================
+
+/// 所有职能部门渠道配置列表
+const List<ChannelConfig> departmentChannels = [
+  ChannelConfig(
+    id: 'latest_info',
+    name: '公开信息',
+    description: '信息公开网 — 学校新闻动态',
+    icon: FluentIcons.news,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 60,
+    defaultEnabled: true,
+  ),
+  ChannelConfig(
+    id: 'notice',
+    name: '通知公示',
+    description: '信息公开网 — 通知公告与公示',
+    icon: FluentIcons.megaphone,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 60,
+    defaultEnabled: true,
+  ),
+  ChannelConfig(
+    id: 'jwc',
+    name: '教务处',
+    description: '教务管理通知与公告',
+    icon: FluentIcons.education,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 60,
+    defaultEnabled: true,
+  ),
+  ChannelConfig(
+    id: 'itc',
+    name: '信息技术中心',
+    description: '校园网与信息系统公告',
+    icon: FluentIcons.settings,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 60,
+    defaultEnabled: true,
+  ),
+  ChannelConfig(
+    id: 'sports',
+    name: '体育部',
+    description: '体育活动与赛事通知',
+    icon: FluentIcons.globe,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'security_dept',
+    name: '保卫处',
+    description: '校园安全通知与公告',
+    icon: FluentIcons.lock,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'construction',
+    name: '基建处',
+    description: '基建处建设要闻与通知公告',
+    icon: FluentIcons.home,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'news_center',
+    name: '新闻网',
+    description: '学校新闻与宣传报道',
+    icon: FluentIcons.news,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 60,
+    defaultEnabled: true,
+  ),
+  ChannelConfig(
+    id: 'student_affairs',
+    name: '学生处',
+    description: '学生管理与服务通知',
+    icon: FluentIcons.mail,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 60,
+    defaultEnabled: true,
+  ),
+  ChannelConfig(
+    id: 'logistics_center',
+    name: '后勤服务中心',
+    description: '通知与新闻动态',
+    icon: FluentIcons.home,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'foreign_student_office',
+    name: '外国留学生事务办公室',
+    description: '公告与新闻',
+    icon: FluentIcons.globe,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'intl_exchange_office',
+    name: '国际交流处',
+    description: '各项新闻与通知公告',
+    icon: FluentIcons.globe,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'admissions_office',
+    name: '招生办',
+    description: '招生动态',
+    icon: FluentIcons.people,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'hr_office',
+    name: '人事处',
+    description: '新闻动态、人才招聘与通知公告',
+    icon: FluentIcons.contact,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'research_office',
+    name: '科研处',
+    description: '科研信息、科研公告与科研成果喜讯',
+    icon: FluentIcons.library,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'union',
+    name: '校工会',
+    description: '工会动态、党建引领与公告通知',
+    icon: FluentIcons.people,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'party_org_dept',
+    name: '党委组织部',
+    description: '党建动态与通知公告',
+    icon: FluentIcons.people,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'united_front_dept',
+    name: '党委统战部',
+    description: '工作动态、党派之声与团体风采',
+    icon: FluentIcons.people,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'party_office',
+    name: '党委办公室',
+    description: '工作动态',
+    icon: FluentIcons.home,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'youth_league',
+    name: '校团委',
+    description: '工作要讯、通知公告与基层动态',
+    icon: FluentIcons.people,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'assets_lab_office',
+    name: '资产与实验管理处',
+    description: '部门新闻与通知公告',
+    icon: FluentIcons.settings,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'sspu_news',
+    name: '学校新闻',
+    description: '学校官网 — 学校新闻',
+    icon: FluentIcons.news,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 60,
+    defaultEnabled: true,
+  ),
+  ChannelConfig(
+    id: 'sspu_notice',
+    name: '学校通知公告',
+    description: '学校官网 — 通知公告',
+    icon: FluentIcons.megaphone,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 60,
+    defaultEnabled: true,
+  ),
+  ChannelConfig(
+    id: 'sspu_activity',
+    name: '校内活动',
+    description: '学校官网 — 校内活动',
+    icon: FluentIcons.event,
+    group: ChannelGroup.department,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+];
