+++ lib/models/channel_config_teaching.dart
@@ -1,0 +1,207 @@
+part of 'channel_config.dart';
+
+// ==================== 教学单位渠道 ====================
+
+/// 所有教学单位渠道配置列表
+const List<ChannelConfig> teachingChannels = [
+  ChannelConfig(
+    id: 'college_cs',
+    name: '计算机与信息工程学院',
+    description: '工作动态、教师工作与学生工作',
+    icon: FluentIcons.settings,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'college_im',
+    name: '智能制造与控制工程学院',
+    description: '学院动态、教学科研与通知公告',
+    icon: FluentIcons.settings,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'college_re',
+    name: '资源与环境工程学院',
+    description: '新闻资讯、通知公告、科研与服务与党建思政',
+    icon: FluentIcons.globe,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'college_em',
+    name: '能源与材料学院',
+    description: '新闻资讯、通知与公告、育人园地与科学研究',
+    icon: FluentIcons.settings,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'college_ic',
+    name: '集成电路学院',
+    description: '学院动态、通知公告、学术活动与科研动态',
+    icon: FluentIcons.settings,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'college_imhe',
+    name: '智能医学与健康工程学院',
+    description: '学院新闻与通知公告',
+    icon: FluentIcons.settings,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'college_econ',
+    name: '经济与管理学院',
+    description: '学院动态、通知公告、育人园地与党群引领',
+    icon: FluentIcons.database,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'college_lang',
+    name: '语言与文化传播学院',
+    description: '新闻动态、学院公告、学生活动与讲座信息',
+    icon: FluentIcons.read,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'college_math',
+    name: '数理与统计学院',
+    description: '学院新闻、学院公告、学术动态与育人园地',
+    icon: FluentIcons.education,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'college_art',
+    name: '艺术与设计学院',
+    description: '学院动态',
+    icon: FluentIcons.video,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'college_vte',
+    name: '职业技术教师教育学院',
+    description: '新闻动态与通知公告',
+    icon: FluentIcons.education,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'college_vt',
+    name: '职业技术学院',
+    description: '学院新闻与学院公告',
+    icon: FluentIcons.education,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'college_marx',
+    name: '马克思主义学院',
+    description: '学院新闻、通知公告、学术科研与教育教学',
+    icon: FluentIcons.library,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'college_ce',
+    name: '继续教育学院',
+    description: '学院新闻与学院公告',
+    icon: FluentIcons.education,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'center_art_edu',
+    name: '艺术教育中心',
+    description: '新闻动态与讲座演出',
+    icon: FluentIcons.video,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'center_intl',
+    name: '国际教育中心',
+    description: '新闻与公告',
+    icon: FluentIcons.globe,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'center_innov',
+    name: '创新创业教育中心',
+    description: '双创教育、通知公告、实践竞赛与创业实践',
+    icon: FluentIcons.settings,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'center_training',
+    name: '工程训练与创新教育中心',
+    description: '中心动态与通知公告',
+    icon: FluentIcons.education,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'graduate',
+    name: '研究生处',
+    description: '研究生管理通知与公告',
+    icon: FluentIcons.education,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+  ChannelConfig(
+    id: 'lib_center',
+    name: '图书馆',
+    description: '新闻动态、通知公告与讲座培训',
+    icon: FluentIcons.library,
+    group: ChannelGroup.teaching,
+    implemented: true,
+    defaultInterval: 120,
+    defaultEnabled: false,
+  ),
+];
