+++ lib/services/message_state_service.dart
@@ -12,6 +12,10 @@ import '../models/channel_config.dart';
 import '../models/message_item.dart';
 import '../services/storage_service.dart';
 
+part 'message_state_service_legacy.dart';
+part 'message_state_service_notifications.dart';
+part 'message_state_service_messages.dart';
+
 /// 消息渠道配置键名
 class MessageChannelKeys {
   MessageChannelKeys._();
@@ -67,7 +71,11 @@ class MessageChannelKeys {
 
 /// 消息状态管理服务（单例）
 /// 管理已读/未读状态持久化和渠道开关
-class MessageStateService {
+class MessageStateService
+    with
+        MessageStateServiceLegacyIntervals,
+        MessageStateServiceNotifications,
+        MessageStateServiceMessages {
   MessageStateService._();
 
   static final MessageStateService instance = MessageStateService._();
@@ -383,212 +391,5 @@ class MessageStateService {
       _channelAutoFetchCountKey(channelId),
       count.clamp(1, 200),
     );
-  }
-
-  // ==================== 自动刷新间隔管理（旧接口，保持兼容） ====================
-
-  /// 获取最新公开信息自动刷新间隔（分钟，0 = 关闭，默认 60）
-  Future<int> getLatestInfoInterval() async {
-    return (await StorageService.getInt(
-          MessageChannelKeys.latestInfoInterval,
-        )) ??
-        60;
-  }
-
-  /// 设置最新公开信息自动刷新间隔（分钟）
-  Future<void> setLatestInfoInterval(int minutes) async {
-    await StorageService.setInt(MessageChannelKeys.latestInfoInterval, minutes);
-  }
-
-  /// 获取通知公示自动刷新间隔（分钟，0 = 关闭，默认 60）
-  Future<int> getNoticeInterval() async {
-    return (await StorageService.getInt(MessageChannelKeys.noticeInterval)) ??
-        60;
-  }
-
-  /// 设置通知公示自动刷新间隔（分钟）
-  Future<void> setNoticeInterval(int minutes) async {
-    await StorageService.setInt(MessageChannelKeys.noticeInterval, minutes);
-  }
-
-  /// 获取微信公众号自动刷新间隔（分钟，0 = 关闭，默认 0）
-  Future<int> getWechatPublicInterval() async {
-    return (await StorageService.getInt(
-          MessageChannelKeys.wechatPublicInterval,
-        )) ??
-        0;
-  }
-
-  /// 设置微信公众号自动刷新间隔（分钟）
-  Future<void> setWechatPublicInterval(int minutes) async {
-    await StorageService.setInt(
-      MessageChannelKeys.wechatPublicInterval,
-      minutes,
-    );
-  }
-
-  /// 获取微信服务号自动刷新间隔（分钟，0 = 关闭，默认 0）
-  Future<int> getWechatServiceInterval() async {
-    return (await StorageService.getInt(
-          MessageChannelKeys.wechatServiceInterval,
-        )) ??
-        0;
-  }
-
-  /// 设置微信服务号自动刷新间隔（分钟）
-  Future<void> setWechatServiceInterval(int minutes) async {
-    await StorageService.setInt(
-      MessageChannelKeys.wechatServiceInterval,
-      minutes,
-    );
-  }
-
-  // ==================== 消息推送与勿扰模式 ====================
-
-  /// 获取消息推送全局开关（默认开启）
-  Future<bool> isNotificationEnabled() async {
-    return await StorageService.getBool(
-      MessageChannelKeys.notificationEnabled,
-      defaultValue: true,
-    );
-  }
-
-  /// 设置消息推送全局开关
-  Future<void> setNotificationEnabled(bool enabled) async {
-    await StorageService.setBool(
-      MessageChannelKeys.notificationEnabled,
-      enabled,
-    );
-  }
-
-  /// 获取勿扰模式是否开启（默认关闭）
-  Future<bool> isDndEnabled() async {
-    return await StorageService.getBool(MessageChannelKeys.dndEnabled);
-  }
-
-  /// 设置勿扰模式开关
-  Future<void> setDndEnabled(bool enabled) async {
-    await StorageService.setBool(MessageChannelKeys.dndEnabled, enabled);
-  }
-
-  /// 获取勿扰开始时间（默认 22:00）
-  Future<int> getDndStartHour() async {
-    return (await StorageService.getInt(MessageChannelKeys.dndStartHour)) ?? 22;
-  }
-
-  /// 获取勿扰开始分钟（默认 0）
-  Future<int> getDndStartMinute() async {
-    return (await StorageService.getInt(MessageChannelKeys.dndStartMinute)) ??
-        0;
-  }
-
-  /// 获取勿扰结束时间（默认 7:00）
-  Future<int> getDndEndHour() async {
-    return (await StorageService.getInt(MessageChannelKeys.dndEndHour)) ?? 7;
-  }
-
-  /// 获取勿扰结束分钟（默认 0）
-  Future<int> getDndEndMinute() async {
-    return (await StorageService.getInt(MessageChannelKeys.dndEndMinute)) ?? 0;
-  }
-
-  /// 设置勿扰时间段（一次性保存开始和结束时间）
-  Future<void> setDndTime({
-    required int startHour,
-    required int startMinute,
-    required int endHour,
-    required int endMinute,
-  }) async {
-    await StorageService.setInt(MessageChannelKeys.dndStartHour, startHour);
-    await StorageService.setInt(MessageChannelKeys.dndStartMinute, startMinute);
-    await StorageService.setInt(MessageChannelKeys.dndEndHour, endHour);
-    await StorageService.setInt(MessageChannelKeys.dndEndMinute, endMinute);
-  }
-
-  /// 判断当前时间是否在勿扰时段内
-  /// 支持跨午夜时段（如 22:00–7:00）
-  Future<bool> isInDndPeriod() async {
-    final enabled = await isDndEnabled();
-    if (!enabled) return false;
-
-    final startH = await getDndStartHour();
-    final startM = await getDndStartMinute();
-    final endH = await getDndEndHour();
-    final endM = await getDndEndMinute();
-
-    final now = DateTime.now();
-    // 将时间转为当天分钟数以便比较
-    final nowMinutes = now.hour * 60 + now.minute;
-    final startMinutes = startH * 60 + startM;
-    final endMinutes = endH * 60 + endM;
-
-    if (startMinutes <= endMinutes) {
-      // 同天时段，如 8:00–12:00
-      return nowMinutes >= startMinutes && nowMinutes < endMinutes;
-    } else {
-      // 跨午夜时段，如 22:00–7:00
-      return nowMinutes >= startMinutes || nowMinutes < endMinutes;
-    }
-  }
-
-  // ==================== 消息持久化管理 ====================
-
-  /// 保存消息列表到本地存储
-  Future<void> saveMessages(List<MessageItem> messages) async {
-    final jsonList = messages.map((msg) => msg.toJson()).toList();
-    await StorageService.setString(
-      MessageChannelKeys.persistedMessages,
-      jsonEncode(jsonList),
-    );
-  }
-
-  /// 从本地存储加载消息列表
-  Future<List<MessageItem>> loadMessages() async {
-    final stored = await StorageService.getString(
-      MessageChannelKeys.persistedMessages,
-    );
-    if (stored == null || stored.isEmpty) return [];
-    try {
-      final jsonList = jsonDecode(stored) as List<dynamic>;
-      return jsonList
-          .map((json) => MessageItem.fromJson(json as Map<String, dynamic>))
-          .toList();
-    } catch (_) {
-      // 存储数据损坏时返回空列表
-      return [];
-    }
-  }
-
-  /// 清除所有微信公众号类型的预存文章
-  /// 保留其他渠道的消息不受影响
-  /// :return: 被清除的文章数量
-  Future<int> clearWechatArticles() async {
-    final messages = await loadMessages();
-    final before = messages.length;
-    messages.removeWhere(
-      (msg) => msg.sourceType == MessageSourceType.wechatPublic,
-    );
-    await saveMessages(messages);
-    return before - messages.length;
-  }
-
-  /// 将新抓取的消息与已有消息合并（按 ID 去重）
-  /// [existingMessages] 已有消息列表
-  /// [newMessages] 新抓取的消息列表
-  /// 返回合并后的列表（已存消息保持原样，避免刷新覆盖原始时间）
-  List<MessageItem> mergeMessages(
-    List<MessageItem> existingMessages,
-    List<MessageItem> newMessages,
-  ) {
-    final messageMap = <String, MessageItem>{};
-    // 先放旧消息，保留首次获取时记录的时间与字段。
-    for (final msg in existingMessages) {
-      messageMap[msg.id] = msg;
-    }
-    // 只补充新增消息，已有文章不再被新抓取结果覆盖。
-    for (final msg in newMessages) {
-      messageMap.putIfAbsent(msg.id, () => msg);
-    }
-    return messageMap.values.toList();
   }
 }
